Fixed instability - new working commit
Tested to run stable for about an hour. Issue was trying to store a
32-bit number in a 16-bit variable, added check to only run tasks once
within a tick.

+++ lib/richard/schedule.cpp
@@ -1,7 +1,7 @@
 #include "schedule.h"
 #include "tasks.h"
 
-Task::Task(uint16_t timeout, uint16_t Interval){
+Task::Task(uint16_t timeout, uint32_t Interval){
   timeOut = timeout;
   lastRun = 0;
   runStart = 0;
@@ -10,11 +10,11 @@ Task::Task(uint16_t timeout, uint16_t Interval){
 }
 
 
-void Task::setLastRun(uint16_t lastrun){
+void Task::setLastRun(uint32_t lastrun){
   lastRun = lastrun;
 }
 
- void Task::setRunStart(uint16_t runstart){
+ void Task::setRunStart(uint32_t runstart){
   runStart = runstart;
 }
